# Conflicts:
#	lib/data/local/database.dart

+++ lib/data/local/database.dart
@@ -117,7 +117,7 @@ class SQLHelper {
   //////////////////////////////////////////////////////////////////////////////
 
 //add
-  static Future<int> addInvoice(int price, List<Product> products , int gain) async {
+  static Future<int> addInvoice(double price, List<Product> products , double gain) async {
     final db = await SQLHelper.initDb(); //open database
     final json = jsonEncode(products);
 
@@ -131,7 +131,7 @@ class SQLHelper {
     var hour = d12.substring(12 , 23);
 
     final data = {
-      'price': price,
+      'price': price as int,
       'products': json,
       'time': currentTime,
       'gain' : gain,
